Refactor create config file
chmod doesn't work for windows operating system. Hence, in windows
don't create default config file and check for permission. This will
let users responsible for creating their own config file and set
appropriate permission.
